+++ src/test/java/net/sf/mpxj/junit/CustomerDataTest.java
@@ -643,7 +643,7 @@ public class CustomerDataTest
 
    private static final Date BASELINE_CURRENT_DATE = new Date(1544100702438L);
 
-   private static final boolean DEBUG_FAILURES = false;
+   private static final boolean DEBUG_FAILURES = true;
 
    static
    {